+++ libsrc/_DEVELOPMENT/math/float/math32/c/m32_math.h
@@ -107,7 +107,7 @@ float m32_sqrf (float a) __z88dk_fastcall;
 float m32_invf (float a) __z88dk_fastcall;
 float m32_sqrtf (float a) __z88dk_fastcall;
 float m32_invsqrtf (float a) __z88dk_fastcall;
-float m32_frexpf (float x, int8_t *pw2) __z88dk_callee;
+float m32_frexpf (float x, int16_t *pw2) __z88dk_callee;
 float m32_ldexpf (float x, int16_t pw2) __z88dk_callee;
 float m32_hypotf (float x, float y) __z88dk_callee;
 float m32_polyf (const float x, const float d[], uint16_t n) __z88dk_callee;
